Last Update

+++ src/main/java/com/global/classes/StudentDTO.java
@@ -28,7 +28,7 @@ public class StudentDTO {
 	}
 
 
-	public float getId() {
+	public int getId() {
 		return id;
 	}
 	public void setId(int id) {
